Parameterize strawberry tests
