+++ tests/framework_strawberry/conftest.py
@@ -40,12 +40,5 @@ collector_agent_registration = collector_agent_registration_fixture(
 )
 
 
-@pytest.fixture(scope="session")
-def app():
-    from _target_application import _target_application
-
-    return _target_application
-
-
 if six.PY2:
     collect_ignore = ["test_application_async.py"]
